% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_params}
\alias{check_params}
\title{Check parameters}
\usage{
check_params(fun_name, parameters)
}
\arguments{
\item{fun_name}{\code{character} string; the name of the function.}

\item{parameters}{\code{character} vector; acceptable parameter names.}
}
\description{
This functions checks if the parameters of a given function
belong to the pre-defined parameter list.
}
\examples{
f1 <- function(x = 1)
  x

check_params("f1", "x")
}
